# backend/tests/contract/openapi.json
{
  "components": {
    "schemas": {
      "AdminSessionRequest": {
        "properties": {
          "password": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "AdminSessionRequest",
        "type": "object"
      },
...
      "DisputeDetail": {
        "properties": {
          "case_id": {
            "format": "uuid",
            "title": "Case Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "customer_id": {
            "title": "Customer Id",
            "type": "string"
          },
          "events": {
            "description": "Append-only timeline: created, then resolved or escalated.",
            "items": {
              "$ref": "#/components/schemas/DisputeEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "evidence": {
            "additionalProperties": true,
            "description": "The transaction as it was, plus resolution (auto_resolved) or handoff (escalated).",
            "title": "Evidence",
            "type": "object"
          },
          "reason_code": {
            "title": "Reason Code",
            "type": "string"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "transaction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null for an escalation that could not be tied to a transaction.",
            "title": "Transaction Id"
          }
        },
        "required": [
          "case_id",
          "customer_id",
          "reason_code",
          "summary",
          "status",
          "evidence",
          "created_at",
          "events"
        ],
        "title": "DisputeDetail",
        "type": "object"
      },
      "DisputeEvent": {
        "properties": {
          "event": {
            "title": "Event",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          }
        },
        "required": [
          "event",
          "payload",
          "ts"
        ],
        "title": "DisputeEvent",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SessionRequest": {
        "properties": {
          "customer_id": {
            "minLength": 1,
            "title": "Customer Id",
            "type": "string"
          },
          "document_number": {
            "minLength": 1,
            "title": "Document Number",
            "type": "string"
          }
        },
        "required": [
          "customer_id",
          "document_number"
        ],
        "title": "SessionRequest",
        "type": "object"
      },
      "SessionResponse": {
        "properties": {
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "session_token": {
            "title": "Session Token",
            "type": "string"
          },
          "token_type": {
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "session_token",
          "token_type",
          "expires_in"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
      "TransitionRequest": {
        "properties": {
          "action": {
            "pattern": "^(claim|close)$",
            "title": "Action",
            "type": "string"
          },
          "note": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Note",
            "type": "string"
          },
          "resolution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required when action=close: resolved_customer or no_resolution",
            "title": "Resolution"
          }
        },
        "required": [
          "action",
          "note"
        ],
        "title": "TransitionRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "\nMock banking service (tool layer) for the Factored AI & Data Hackathon 2026\nprototype. Sandbox limitations, stated explicitly:\n\n- Login is a test session: `POST /session` checks `customer_id` +\n  `document_number` against `gold.customers` and issues a short-lived JWT\n  (HS256). No real identity provider is behind it.\n- Admin login (`POST /admin/session`) checks credentials against `ADMIN_USERS`\n  (bcrypt hashes, in-memory rate limit, single-process sandbox limitation).\n  Admins work an inbox of escalated cases with audited transitions\n  (claim -> in_progress -> closed); they never set `auto_resolved`.\n- `GET /meta/demo-scenarios` is a public demo aid: it surfaces only the fields\n  the test login itself asks for (customer_id, document_number, name), never\n  fraud columns.\n- Reads come from the read-only `gold.*` tables published by the dbt pipeline;\n  operational tables (`app.*`) are built by the versioned SQL migrations in\n  `app/migrations`, applied at startup under an advisory lock.\n- `/v1` is what the web app uses (user name and password login, products,\n  paginated history, summary). The routes at the root are the agent's.\n- No money is ever moved: disputes only explain, document and escalate.\n- Fraud ground truth (`is_fraud`/`fraud_score`) is not part of `gold.transactions`\n  and is never exposed here.\n",
    "title": "mock-banking-service",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/disputes": {
      "get": {
        "operationId": "list_disputes_admin_disputes_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "customer_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 2,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Customer Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Disputes Admin Disputes Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Disputes",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/disputes/{case_id}": {
      "get": {
        "operationId": "get_dispute_admin_disputes__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Dispute Admin Disputes  Case Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Dispute",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/disputes/{case_id}/transition": {
      "post": {
        "operationId": "transition_dispute_admin_disputes__case_id__transition_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Transition Dispute Admin Disputes  Case Id  Transition Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Transition Dispute",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/admin/session": {
      "post": {
        "operationId": "create_admin_session_admin_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Admin Session",
        "tags": [
          "session",
          "admin"
        ]
      }
    },
...
    "/disputes/{case_id}": {
      "get": {
        "operationId": "get_dispute_disputes__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Dispute",
        "tags": [
          "disputes"
        ]
      }
    },
...
    "/me/disputes": {
      "get": {
        "operationId": "list_my_disputes_me_disputes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List My Disputes Me Disputes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List My Disputes",
        "tags": [
          "customers"
        ]
      }
    },
...
    "/meta/data": {
      "get": {
        "description": "Data freshness for the admin console: gold.* counts plus the last pipeline run\n(ops.etl_runs). The snapshot is static by design — see spec/ARCHITECTURE.md.",
        "operationId": "meta_data_meta_data_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Meta Data Meta Data Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Meta Data",
        "tags": [
          "meta"
        ]
      }
    },
...
    "/session": {
      "post": {
        "operationId": "create_session_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session",
        "tags": [
          "session"
        ]
      }
    },
...
    "/v1/disputes/{case_id}": {
      "get": {
        "operationId": "get_dispute_v1_disputes__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Dispute",
        "tags": [
          "disputes"
        ]
      }
    },
...
  }
}